+++ webext/lib/settings.js
@@ -1,5 +1,7 @@
 class Settings {
     constructor() {
         this.attestedSATDomainsOnly = false;
+        // Rollout relaxtaions
+        this.wildcardSATDomainsAllowed = false;
     }
 }