+++ webext/lib/settings.js
@@ -3,5 +3,6 @@ class Settings {
         this.attestedSATDomainsOnly = false;
         // Rollout relaxtaions
         this.wildcardSATDomainsAllowed = false;
+        this.satAltSvcNotInTLSCertAllowed = false;
     }
 }