+++ webext/lib/log.js
@@ -12,11 +12,23 @@ function log_assert(cond, msg) {
     return cond;
 }
 
-/**
- * Log something to the console
- */
-function log() {
-    return console.log("[SATIS]", [].slice.apply(arguments).join(" "));
+function __dateStr() {
+    let t = new Date();
+    let h = t.getHours();
+    let m = t.getMinutes();
+    let s = t.getSeconds();
+    let ms = t.getMilliseconds();
+    h = "" + h;
+    m = m < 10 ? "0" + m : "" + m;
+    s = s < 10 ? "0" + s : "" + s;
+    ms = ms < 100 ? ms < 10 ? "00" + ms : "0" + ms : "" + ms;
+    return h + ":" + m + ":" + s + "." + ms;
+}
+
+function __log(level) {
+    return console.log(
+        "[SAT]", "[" + level + "]", "[" + __dateStr() + "]",
+        [].slice.apply(arguments).slice(1).join(" "));
 }
 
 function log_object(o) {
@@ -26,13 +38,13 @@ function log_object(o) {
 
 function log_debug() {
     if (LOG_DEBUG)
-        return log("[DEBUG]", [].slice.apply(arguments).join(" "));
+        return __log("DEBUG", [].slice.apply(arguments).join(" "));
 }
 
 function log_warn() {
-    return log("[WARN]", [].slice.apply(arguments).join(" "));
+    return __log("WARN", [].slice.apply(arguments).join(" "));
 }
 
 function log_error() {
-    return log("[ERROR]", [].slice.apply(arguments).join(" "));
+    return __log("ERROR", [].slice.apply(arguments).join(" "));
 }