+++ webext/CHANGELOG.md
@@ -27,6 +27,7 @@ domain) to be in the TLS certificate too is *more* strict.
 - Catch exception when trying to get securityInfo and the API doesn't exist.
 - Store and display all AltSvc headers we get, not just the SAT and onion ones
 - Min version of FF only needs to be 63, not 65
+- Improve log strings with the time
 
 ## v1.0.2 - 2019-02-19